requirements.txt
flask==3.1.1
google==3.0.0
matplotlib==3.10.1
networkx==3.5
numba==0.61.0
numpy==1.26.4
openai==1.82.0
pandas==2.3.1
psutil==5.9.0
pymoo==0.6.1.5
pytz==2025.2
requests==2.32.3
scipy==1.15.3
setuptools==75.8.0
sympy==1.14.0
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
python-dotenv
tree_sitter